Made SerializationFeatureHelper obsolete - handled by th ConfigureSerialization Feature class
