+++ src/NServiceBus.Core/Serialization/ConfigureSerialization.cs
@@ -1,6 +1,8 @@
 ﻿namespace NServiceBus.Serialization
 {
     using System;
+    using System.Collections.Generic;
+    using System.Linq;
     using NServiceBus.Features;
     using NServiceBus.MessageInterfaces.MessageMapper.Reflection;
     using NServiceBus.Serializers;
@@ -56,6 +58,27 @@
 
             var c = context.Container.ConfigureComponent(serializerType, DependencyLifecycle.SingleInstance);
             context.Settings.ApplyTo(serializerType, c);
+        }
+
+        private bool IsDefaultSerializer(FeatureConfigurationContext context)
+        {
+            Guard.AgainstNull("context", context);
+
+            var serializationDefinition = context.Settings.GetSelectedSerializer();
+            return serializationDefinition.ProvidedByFeature() == GetType();
+        }
+
+        private bool IsAdditionalDeserializer(FeatureConfigurationContext context)
+        {
+            Guard.AgainstNull("context", context);
+
+            HashSet<SerializationDefinition> deserializers;
+            if (!context.Settings.TryGet("AdditionalDeserializers", out deserializers))
+            {
+                return false;
+            }
+
+            return deserializers.Any(definition => definition.ProvidedByFeature() == GetType());
         }
     }
 }
